+++ src/__tests__/index.test.js
@@ -294,4 +294,65 @@ describe('denormalize', () => {
 
     expect(denormalize(normalizedData.result, [patronsSchema], normalizedData.entities)).toMatchSnapshot();
   });
+
+  test('denormalize with sub entity delete key set to truthy value', () => {
+    const user = new schema.Entity('users');
+    const article = new schema.Entity('articles', { author: user });
+    const entities = {
+      articles: {
+        '123': {
+          id: '123',
+          title: 'A Great Article',
+          author: '8472'
+        }
+      },
+      users: {
+        '8472': {
+          id: '8472',
+          name: 'Paul',
+          [user.deleteKey]: true
+        }
+      }
+    };
+    expect(denormalize('123', article, entities)).toEqual({
+      id: '123',
+      title: 'A Great Article',
+      author: null
+    });
+  });
+
+  test('denormalize with sub entity delete key set to true', () => {
+    const user = new schema.Entity('users');
+    const article = new schema.Entity('articles', { authors: [user] });
+    const entities = {
+      articles: {
+        '123': {
+          id: '123',
+          title: 'A Great Article',
+          authors: ['8472', '8471']
+        }
+      },
+      users: {
+        '8471': {
+          id: '8471',
+          name: 'Frank'
+        },
+        '8472': {
+          id: '8472',
+          name: 'Paul',
+          [user.deleteKey]: true
+        }
+      }
+    };
+    expect(denormalize('123', article, entities)).toEqual({
+      id: '123',
+      title: 'A Great Article',
+      authors: [
+        {
+          id: '8471',
+          name: 'Frank'
+        }
+      ]
+    });
+  });
 });