+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v3.5.3
+
+* **Fixed/Changed** 2th/3rd params of Entity.getId become optional
+
+# v3.5.2
+
+* **Added** Support deleteKey for `EntitySchema`
+
 # v3.4.0
 
 * **Changed** Now built with Babel 7